refactor main screen to utilize BookingItem model, streamline token management, and enhance UI with loading states
