+++ lib/features/main_screen/renter/presentation/views/main_screen_body.dart
@@ -1,0 +1,250 @@
+import 'package:aggar/core/extensions/context_colors_extension.dart';
+import 'package:aggar/features/main_screen/renter/model/booking_item.dart';
+import 'package:aggar/features/main_screen/renter/presentation/widgets/booking_card.dart';
+import 'package:aggar/features/main_screen/widgets/main_header.dart';
+import 'package:aggar/features/notification/presentation/views/notification_screen.dart';
+import 'package:flutter/material.dart';
+import 'package:table_calendar/table_calendar.dart';
+import 'package:gap/gap.dart';
+
+class MainScreenBody extends StatefulWidget {
+  const MainScreenBody({
+    super.key,
+    required this.accessToken,
+    required this.onRefresh,
+  });
+
+  final String accessToken;
+  final VoidCallback onRefresh;
+
+  @override
+  State<MainScreenBody> createState() => _MainScreenBodyState();
+}
+
+class _MainScreenBodyState extends State<MainScreenBody> {
+  CalendarFormat _calendarFormat = CalendarFormat.month;
+  DateTime _focusedDay = DateTime.now();
+  DateTime? _selectedDay;
+
+  final List<BookingItem> _bookings = [
+    BookingItem(
+      name: 'Mike Chen',
+      service: 'Hair Cut & Trim',
+      time: '2:15 PM',
+      date: 'Dec 10',
+      status: 'Pending',
+      color: const Color(0xFF6B73FF),
+      initial: 'M',
+    ),
+    BookingItem(
+      name: 'Emma Davis',
+      service: 'Color Treatment',
+      time: '4:00 PM',
+      date: 'Dec 10',
+      status: 'Pending',
+      color: const Color(0xFF8B5CF6),
+      initial: 'E',
+    ),
+  ];
+
+  @override
+  void initState() {
+    super.initState();
+    _selectedDay = DateTime.now();
+  }
+
+  Future<void> _onRefresh() async {
+    widget.onRefresh();
+    await Future.delayed(const Duration(seconds: 1));
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return RefreshIndicator(
+      onRefresh: _onRefresh,
+      child: SingleChildScrollView(
+        physics: const AlwaysScrollableScrollPhysics(),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: [
+            Container(
+              decoration: BoxDecoration(
+                color: context.theme.blue100_8,
+                borderRadius: const BorderRadius.only(
+                  bottomLeft: Radius.circular(20),
+                  bottomRight: Radius.circular(20),
+                ),
+              ),
+              padding: const EdgeInsets.only(
+                  left: 20, right: 20, top: 55, bottom: 20),
+              child: MainHeader(
+                onPressed: () => Navigator.push(
+                    context,
+                    MaterialPageRoute(
+                      builder: (context) => const NotificationScreen(),
+                    )),
+                accesstoken: widget.accessToken,
+              ),
+            ),
+
+            // Calendar Section
+            Padding(
+              padding: const EdgeInsets.all(10.0),
+              child: Container(
+                decoration: BoxDecoration(
+                  color: Colors.white,
+                  borderRadius: BorderRadius.circular(16),
+                  boxShadow: [
+                    BoxShadow(
+                      color: Colors.grey.withOpacity(0.1),
+                      spreadRadius: 1,
+                      blurRadius: 10,
+                      offset: const Offset(0, 2),
+                    ),
+                  ],
+                ),
+                child: Padding(
+                  padding: const EdgeInsets.all(15.0),
+                  child: Column(
+                    crossAxisAlignment: CrossAxisAlignment.start,
+                    children: [
+                      const Text(
+                        'Calendar',
+                        style: TextStyle(
+                          fontSize: 18,
+                          fontWeight: FontWeight.w600,
+                          color: Colors.black87,
+                        ),
+                      ),
+                      const Gap(16),
+                      TableCalendar<dynamic>(
+                        firstDay: DateTime.utc(2020, 1, 1),
+                        lastDay: DateTime.utc(2030, 12, 31),
+                        focusedDay: _focusedDay,
+                        calendarFormat: _calendarFormat,
+                        selectedDayPredicate: (day) {
+                          return isSameDay(_selectedDay, day);
+                        },
+                        onDaySelected: (selectedDay, focusedDay) {
+                          if (!isSameDay(_selectedDay, selectedDay)) {
+                            setState(() {
+                              _selectedDay = selectedDay;
+                              _focusedDay = focusedDay;
+                            });
+                          }
+                        },
+                        onFormatChanged: (format) {
+                          if (_calendarFormat != format) {
+                            setState(() {
+                              _calendarFormat = format;
+                            });
+                          }
+                        },
+                        onPageChanged: (focusedDay) {
+                          _focusedDay = focusedDay;
+                        },
+                        headerStyle: const HeaderStyle(
+                          formatButtonVisible: false,
+                          titleCentered: true,
+                          leftChevronIcon: Icon(
+                            Icons.chevron_left,
+                            color: Colors.grey,
+                          ),
+                          rightChevronIcon: Icon(
+                            Icons.chevron_right,
+                            color: Colors.grey,
+                          ),
+                          titleTextStyle: TextStyle(
+                            fontSize: 16,
+                            fontWeight: FontWeight.w600,
+                            color: Colors.black87,
+                          ),
+                        ),
+                        calendarStyle: CalendarStyle(
+                          outsideDaysVisible: false,
+                          weekendTextStyle: const TextStyle(
+                            color: Colors.black87,
+                          ),
+                          holidayTextStyle: const TextStyle(
+                            color: Colors.black87,
+                          ),
+                          selectedDecoration: const BoxDecoration(
+                            color: Color(0xFF6B73FF),
+                            shape: BoxShape.circle,
+                          ),
+                          todayDecoration: BoxDecoration(
+                            color: Colors.grey[300],
+                            shape: BoxShape.circle,
+                          ),
+                          defaultTextStyle: const TextStyle(
+                            color: Colors.black87,
+                          ),
+                        ),
+                        daysOfWeekStyle: const DaysOfWeekStyle(
+                          weekdayStyle: TextStyle(
+                            color: Colors.grey,
+                            fontWeight: FontWeight.w500,
+                          ),
+                          weekendStyle: TextStyle(
+                            color: Colors.grey,
+                            fontWeight: FontWeight.w500,
+                          ),
+                        ),
+                      ),
+                    ],
+                  ),
+                ),
+              ),
+            ),
+
+            const Gap(24),
+
+            Container(
+              decoration: BoxDecoration(
+                color: Colors.white,
+                borderRadius: BorderRadius.circular(16),
+                boxShadow: [
+                  BoxShadow(
+                    color: Colors.grey.withOpacity(0.1),
+                    spreadRadius: 1,
+                    blurRadius: 10,
+                    offset: const Offset(0, 2),
+                  ),
+                ],
+              ),
+              child: Padding(
+                padding: const EdgeInsets.all(16.0),
+                child: Column(
+                  crossAxisAlignment: CrossAxisAlignment.start,
+                  children: [
+                    const Text(
+                      'New Bookings',
+                      style: TextStyle(
+                        fontSize: 18,
+                        fontWeight: FontWeight.w600,
+                        color: Colors.black87,
+                      ),
+                    ),
+                    const Gap(5),
+                    ListView.separated(
+                      shrinkWrap: true,
+                      physics: const NeverScrollableScrollPhysics(),
+                      itemCount: _bookings.length,
+                      separatorBuilder: (context, index) => const Gap(12),
+                      itemBuilder: (context, index) {
+                        final booking = _bookings[index];
+                        return BookingCard(booking: booking);
+                      },
+                    ),
+                  ],
+                ),
+              ),
+            ),
+
+            const Gap(20),
+          ],
+        ),
+      ),
+    );
+  }
+}
